Fix incorrect ordering constraint
Binary search requires elements to have a Total Ordering, the initial
implementation used only Partial Ordering as this allowed the API
surface to be closer to the original Python library (e.g. one can pass
a `Vec<f32>` to the same function as a `Vec<i32>`). However, upon
further consideration this is **simply incorrect**.

Signed-off-by: Ben Steadman <steadmanben1@gmail.com>