Add bisect with custom comparator functionality
There's a lot of boilerplate code here which could probably be reduced
by introducing a struct of some sort, however I like the idea of keeping
the API surface as standalone functions akin to the original Python
module. Since this code is unlikely to change much (e.g. it's a small,
very limited scope problem), having some additional boilerplate is not
a problem in my opinion.

Signed-off-by: Ben Steadman <steadmanben1@gmail.com>